Delta Robot Kinematic 3D Simulation in OpenGL C#
In this study, the materials coming from the conveyor line will be taken from the conveyor line using delta robot arm and placed in the desired position as required by the process. The program that will perform the 3D simulation of this system will be implemented. Thus, the instant movements of the system can be followed by the user through this simulation program.